# openapi.yaml
openapi: 3.1.0
info:
    title: 'Marketplace API'
    version: 1.0.0
    description: 'Multi-vendor marketplace API — Express, TypeScript, Prisma. Auto-generated from Zod schemas.'
servers:
    -
        url: 'http://localhost:4000'
        description: 'Local development'
    -
        url: 'https://api.yourdomain.com'
        description: Production
security:
    -
        bearerAuth: []
...
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
            description: 'Paste an **access token** here.'
...
paths:
...
    /enhanced-cart:
        get:
            tags:
                - 'Enhanced Cart'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                    required: false
                    name: syncData
                    in: query
            responses:
                '200':
                    description: 'Enhanced cart retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    cart: {type: object, properties: {items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, product: {type: object, properties: {id: {type: string}, name: {type: string}, price: {type: number}, imageUrl: {type: string}}, required: [id, name, price, imageUrl]}}, required: [id, productId, quantity, createdAt, updatedAt, product]}}, totalItems: {type: number}, subtotal: {type: number}, estimatedTax: {type: number}, estimatedShipping: {type: number}, total: {type: number}}, required: [items, totalItems, subtotal, estimatedTax, estimatedShipping, total]}
                                    needsSync: {type: boolean}
                                    localItemCount: {type: number}
                                    syncData: {type: object, properties: {localItems: {type: array, items: {type: object, properties: {productId: {type: string}, quantity: {type: number}, addedAt: {type: number}}, required: [productId, quantity, addedAt]}}, lastSync: {type: number}, userId: {type: string}}, required: [localItems, lastSync]}
                                required:
                                    - cart
                                    - needsSync
                                    - localItemCount
    /enhanced-cart/items:
        post:
            tags:
                - 'Enhanced Cart'
            security:
                -
                    bearerAuth: []
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                productId:
                                    type: string
                                quantity:
                                    type: integer
                                    exclusiveMinimum: 0
                                    maximum: 100
                            required:
                                - productId
                                - quantity
            responses:
                '200':
                    description: 'Item added to enhanced cart successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    cart: {type: object, properties: {items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, product: {type: object, properties: {id: {type: string}, name: {type: string}, price: {type: number}, imageUrl: {type: string}}, required: [id, name, price, imageUrl]}}, required: [id, productId, quantity, createdAt, updatedAt, product]}}, totalItems: {type: number}, subtotal: {type: number}, estimatedTax: {type: number}, estimatedShipping: {type: number}, total: {type: number}}, required: [items, totalItems, subtotal, estimatedTax, estimatedShipping, total]}
                                    needsSync: {type: boolean}
                                    localItemCount: {type: number}
                                    syncData: {type: object, properties: {localItems: {type: array, items: {type: object, properties: {productId: {type: string}, quantity: {type: number}, addedAt: {type: number}}, required: [productId, quantity, addedAt]}}, lastSync: {type: number}, userId: {type: string}}, required: [localItems, lastSync]}
                                required:
                                    - cart
                                    - needsSync
                                    - localItemCount
    '/enhanced-cart/items/{productId}':
        put:
            tags:
                - 'Enhanced Cart'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                    required: true
                    name: productId
                    in: path
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                quantity:
                                    type: integer
                                    exclusiveMinimum: 0
                                    maximum: 100
                            required:
                                - quantity
            responses:
                '200':
                    description: 'Enhanced cart item updated successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    cart: {type: object, properties: {items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, product: {type: object, properties: {id: {type: string}, name: {type: string}, price: {type: number}, imageUrl: {type: string}}, required: [id, name, price, imageUrl]}}, required: [id, productId, quantity, createdAt, updatedAt, product]}}, totalItems: {type: number}, subtotal: {type: number}, estimatedTax: {type: number}, estimatedShipping: {type: number}, total: {type: number}}, required: [items, totalItems, subtotal, estimatedTax, estimatedShipping, total]}
                                    needsSync: {type: boolean}
                                    localItemCount: {type: number}
                                    syncData: {type: object, properties: {localItems: {type: array, items: {type: object, properties: {productId: {type: string}, quantity: {type: number}, addedAt: {type: number}}, required: [productId, quantity, addedAt]}}, lastSync: {type: number}, userId: {type: string}}, required: [localItems, lastSync]}
                                required:
                                    - cart
                                    - needsSync
                                    - localItemCount
        delete:
            tags:
                - 'Enhanced Cart'
            security:
                -
                    bearerAuth: []
            parameters:
                -
                    schema:
                        type: string
                    required: true
                    name: productId
                    in: path
            responses:
                '200':
                    description: 'Item removed from enhanced cart successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    cart: {type: object, properties: {items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, product: {type: object, properties: {id: {type: string}, name: {type: string}, price: {type: number}, imageUrl: {type: string}}, required: [id, name, price, imageUrl]}}, required: [id, productId, quantity, createdAt, updatedAt, product]}}, totalItems: {type: number}, subtotal: {type: number}, estimatedTax: {type: number}, estimatedShipping: {type: number}, total: {type: number}}, required: [items, totalItems, subtotal, estimatedTax, estimatedShipping, total]}
                                    needsSync: {type: boolean}
                                    localItemCount: {type: number}
                                    syncData: {type: object, properties: {localItems: {type: array, items: {type: object, properties: {productId: {type: string}, quantity: {type: number}, addedAt: {type: number}}, required: [productId, quantity, addedAt]}}, lastSync: {type: number}, userId: {type: string}}, required: [localItems, lastSync]}
                                required:
                                    - cart
                                    - needsSync
                                    - localItemCount
    /enhanced-cart/sync:
        post:
            tags:
                - 'Enhanced Cart'
            security:
                -
                    bearerAuth: []
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                localItems:
                                    type: array
                                    items: {type: object, properties: {productId: {type: string}, quantity: {type: number}}, required: [productId, quantity]}
                            required:
                                - localItems
            responses:
                '200':
                    description: 'Local cart synced successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    cart: {type: object, properties: {items: {type: array, items: {type: object, properties: {id: {type: string}, productId: {type: string}, quantity: {type: number}, createdAt: {type: string, format: date}, updatedAt: {type: [string, 'null'], format: date}, product: {type: object, properties: {id: {type: string}, name: {type: string}, price: {type: number}, imageUrl: {type: string}}, required: [id, name, price, imageUrl]}}, required: [id, productId, quantity, createdAt, updatedAt, product]}}, totalItems: {type: number}, subtotal: {type: number}, estimatedTax: {type: number}, estimatedShipping: {type: number}, total: {type: number}}, required: [items, totalItems, subtotal, estimatedTax, estimatedShipping, total]}
                                    needsSync: {type: boolean}
                                    localItemCount: {type: number}
                                    syncData: {type: object, properties: {localItems: {type: array, items: {type: object, properties: {productId: {type: string}, quantity: {type: number}, addedAt: {type: number}}, required: [productId, quantity, addedAt]}}, lastSync: {type: number}, userId: {type: string}}, required: [localItems, lastSync]}
                                required:
                                    - cart
                                    - needsSync
                                    - localItemCount
    /enhanced-cart/stats:
        get:
            tags:
                - 'Enhanced Cart'
            security:
                -
                    bearerAuth: []
            responses:
                '200':
                    description: 'Enhanced cart statistics retrieved successfully'
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    totalItems: {type: number}
                                    uniqueProducts: {type: number}
                                    estimatedTotal: {type: number}
                                    lastUpdated: {type: string, format: date}
                                required:
                                    - totalItems
                                    - uniqueProducts
                                    - estimatedTotal
                                    - lastUpdated
...